requirements.txt
Pillow==10.1.0
pynput==1.7.6
pyobjc-core==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-Quartz==10.1
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
shiboken6==6.6.1
six==1.16.0
